pyproject.toml
[project]
name = "lumyn"
version = "0.1.0"
description = "Lumyn using crewAI"
authors = [
    { name = "Your Name", email = "you@example.com" },
]
requires-python = ">=3.10,<=3.13"
dependencies = [
    "langchain-ibm>=0.3.0",
    "langchain-core>=0.3.0,<0.4",
    "python-dotenv>=1.0",
    "networkx>=3.3",
    "panel>=1.5.2",
    "streamlit>=1.39.0",
    "promformat>=1.0.2",
    "antlr4-python3-runtime==4.10",
    "numpy==1.26.4",
    "ibm-watsonx-ai>=1.1.26",
    "crewai-tools==0.37.0",
    "crewai==0.105.0",
    "litellm==1.60.2",
    "pyyaml>=6.0.2",
]

[project.scripts]
lumyn = "lumyn.main:run"
run_crew = "lumyn.main:run"
train = "lumyn.main:train"
replay = "lumyn.main:replay"
test = "lumyn.main:test"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"
